+++ Controller/Adminhtml/Events/MassEnable.php
@@ -1,18 +1,19 @@
 <?php
 
-namespace Aligent\Webhooks\Controller\Adminhtml\Webhooks;
+namespace Aligent\Webhooks\Controller\Adminhtml\Events;
 
 use Aligent\Webhooks\Api\AsyncEventRepositoryInterface;
 use Aligent\Webhooks\Model\AsyncEvent;
 use Aligent\Webhooks\Model\ResourceModel\Webhook\Collection;
-use Magento\Ui\Component\MassAction\Filter;
 use Aligent\Webhooks\Model\ResourceModel\Webhook\CollectionFactory;
 use Magento\Backend\App\Action;
 use Magento\Backend\App\Action\Context;
 use Magento\Framework\App\Action\HttpPostActionInterface;
+use Magento\Ui\Component\MassAction\Filter;
 
-class MassDisable extends Action implements HttpPostActionInterface
+class MassEnable extends Action implements HttpPostActionInterface
 {
+
     /**
      * @var CollectionFactory
      */
@@ -35,55 +36,52 @@ class MassDisable extends Action implements HttpPostActionInterface
         AsyncEventRepositoryInterface $webhookRepository
     ) {
         parent::__construct($context);
-        $this->collectionFactory = $collectionFactory;
         $this->filter = $filter;
+        $this->collectionFactory = $collectionFactory;
         $this->webhookRepository = $webhookRepository;
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function execute()
     {
         $resultRedirect = $this->resultRedirectFactory->create();
-        $resultRedirect->setPath('webhooks/webhooks/index');
+        $resultRedirect->setPath('async_events/events/index');
 
         $webhookCollection = $this->collectionFactory->create();
         $this->filter->getCollection($webhookCollection);
-        $this->disableWebhooks($webhookCollection);
+        $this->enableWebhooks($webhookCollection);
 
         return $resultRedirect;
     }
 
-    private function disableWebhooks(Collection $webhookCollection)
+    private function enableWebhooks(Collection $webhookCollection)
     {
-        $disabled = 0;
-        $alreadyDisabled = 0;
+        $enabled = 0;
+        $alreadyEnabled = 0;
 
         /** @var AsyncEvent $webhook */
         foreach ($webhookCollection as $webhook) {
-            $alreadyDisabled++;
-            if ($webhook->getStatus()) {
+            $alreadyEnabled++;
+            if (!$webhook->getStatus()) {
                 try {
-                    $webhook->setStatus(false);
+                    $webhook->setStatus(true);
                     $this->webhookRepository->save($webhook, false);
-                    $alreadyDisabled--;
-                    $disabled++;
+                    $alreadyEnabled--;
+                    $enabled++;
                 } catch (\Exception $e) {
                     $this->messageManager->addErrorMessage($e->getMessage());
                 }
             }
         }
 
-        if ($disabled) {
+        if ($enabled) {
             $this->messageManager->addSuccessMessage(
-                __('A total of %1 webhook(s) have been disabled.', $disabled)
+                __('A total of %1 webhook(s) have been enabled.', $enabled)
             );
         }
 
-        if ($alreadyDisabled) {
+        if ($alreadyEnabled) {
             $this->messageManager->addNoticeMessage(
-                __('A total of %1 webhook(s) are already disabled.', $alreadyDisabled)
+                __('A total of %1 webhook(s) are already enabled.', $alreadyEnabled)
             );
         }
     }
