+++ etc/adminhtml/menu.xml
@@ -8,7 +8,7 @@
 
         <add id="Aligent_Webhooks::webhooks" title="Subscribers" translate="title" module="Aligent_Webhooks"
              parent="Aligent_Webhooks::integrations" sortOrder="10" dependsOnModule="Aligent_Webhooks"
-             action="webhooks/webhooks" resource="Aligent_Webhooks::manage" />
+             action="async_events/events" resource="Aligent_Webhooks::manage" />
 
         <add id="Aligent_Webhooks::webhooks_logs" title="Logs" translate="title" module="Aligent_Webhooks"
              parent="Aligent_Webhooks::integrations" sortOrder="10" dependsOnModule="Aligent_Webhooks"
