+++ view/adminhtml/layout/async_events_events_index.xml
@@ -2,7 +2,7 @@
 <page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
     <body>
         <referenceContainer name="content">
-            <uiComponent name="webhooks_webhooks_listing"/>
+            <uiComponent name="async_events_events_listing"/>
         </referenceContainer>
     </body>
 </page>
